requirements.txt
python-dotenv
opencv-python
torch
torchvision
torchaudio
transformers
moviepy
google-generativeai
rich
fpdf
langgraph
langchain
langchain-core
langchain-google-genai
chromadb
numpy
sse-starlette
neo4j
sentence-transformers
cohere
ultralytics
easyocr
pydantic
wget